Add rainbow color effects and settings integration across various layers
- Implemented rainbow color effects in PlayLayer, LevelEditorLayer, MenuLayer, ProfilePage, and GJGarageLayer.
- Introduced a settings structure to manage rainbow effect parameters such as speed, saturation, brightness, and presets.
- Added functionality to apply rainbow colors to player icons in the menu and profile pages.
- Integrated a settings button in the pause menu and editor pause layer for easy access to mod settings.
- Enhanced color application logic to support glow effects and secondary colors based on user-defined presets.
- Organized code into separate files for better maintainability, including settings, color utilities, and color application logic.
