+++ mod.json
@@ -6,7 +6,7 @@
 		"ios": "2.2081",
 		"mac":"2.2081"
 	},
-	"version": "1.4.2",
+	"version": "1.5.0",
 	"id": "saumondeluxe.rainbow_icon",
 	"name": "Rainbow Icon",
 	"developer": "SaumonDeLuxe",
